authenticated routes backend

+++ aulas/server/src/routes/memories.ts
@@ -3,8 +3,13 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get("/memories", async () => {
+    app.addHook("preHandler", async (request) => {
+        await request.jwtVerify();
+    });
+
+    app.get("/memories", async (request) => {
         const memories = await prisma.memory.findMany({
+            where: { userId: request.user.sub },
             orderBy: {
                 createdAt: "asc",
             },
@@ -19,7 +24,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         });
     });
 
-    app.get("/memories/:id", async (request) => {
+    app.get("/memories/:id", async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -31,6 +36,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             },
         });
+
+        if (!memory.isPublic && memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
 
         return memory;
     });
@@ -49,13 +58,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 isPublic,
                 coverUrl,
-                userId: "a9f8977d-0652-45e9-ad90-9274099aea74",
+                userId: request.user.sub,
             },
         });
         return memory;
     });
 
-    app.put("/memories/:id", async (request) => {
+    app.put("/memories/:id", async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -70,7 +79,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-        const memory = await prisma.memory.update({
+        // Validação
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            },
+        });
+
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
+
+        memory = await prisma.memory.update({
             data: {
                 content,
                 coverUrl,
@@ -83,12 +103,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory;
     });
 
-    app.delete("/memories/:id", async (request) => {
+    app.delete("/memories/:id", async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
 
         const { id } = paramsSchema.parse(request.params);
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            },
+        });
+
+        // Validação
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
 
         await prisma.memory.delete({
             where: {
